requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
torch==2.5.1
transformers==4.46.3
safetensors==0.4.5
pillow==11.0.0
python-multipart==0.0.12
